Move MetricsReporter and it's dependencies into //ledger/metrics
