+++ ledger-service/cli-opts/BUILD.bazel
@@ -27,7 +27,7 @@ da_scala_library(
     ],
     deps = [
         "//ledger/ledger-api-common",
-        "//ledger/participant-integration-api",
+        "//ledger/metrics",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
         "@maven//:io_netty_netty_handler",
