+++ ledger-service/http-json-cli/BUILD.bazel
@@ -20,7 +20,7 @@ da_scala_library(
         "//ledger-service/cli-opts",
         "//ledger-service/utils",
         "//ledger/ledger-api-common",
-        "//ledger/participant-integration-api",
+        "//ledger/metrics",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:io_netty_netty_handler",
         "@maven//:org_slf4j_slf4j_api",
