+++ ledger-service/http-json/BUILD.bazel
@@ -66,9 +66,7 @@ hj_scalacopts = lf_scalacopts + [
             "//ledger/ledger-api-auth",
             "//ledger/ledger-api-common",
             "//ledger/ledger-resources",
-            "//ledger/sandbox-common:sandbox-common-{}".format(edition),
             "//ledger/metrics",
-            "//ledger/participant-integration-api",
             "//libs-scala/auth-utils",
             "//libs-scala/concurrent",
             "//libs-scala/contextualized-logging",
