+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/WebsocketEndpoints.scala
@@ -12,12 +12,12 @@ import scalaz.syntax.std.boolean._
 import scalaz.syntax.std.option._
 import scalaz.\/
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.Future
 import EndpointsCompanion._
 import com.daml.http.domain.JwtPayload
 import com.daml.http.util.Logging.{InstanceUUID, RequestID, extendWithRequestIdLogCtx}
 import com.daml.logging.{ContextualizedLogger, LoggingContextOf}
-import com.daml.metrics.{Metrics, Timed}
+import com.daml.metrics.Metrics
 
 object WebsocketEndpoints {
   private[http] val tokenPrefix: String = "jwt.token."
@@ -57,7 +57,6 @@ class WebsocketEndpoints(
   private[this] val logger = ContextualizedLogger.get(getClass)
 
   def transactionWebSocket(implicit
-      ec: ExecutionContext,
       lc: LoggingContextOf[InstanceUUID],
       metrics: Metrics,
   ) = {
@@ -109,14 +108,8 @@ class WebsocketEndpoints(
     import scalaz.std.partialFunction._, scalaz.syntax.arrow._
     (dispatch &&& { case r => r }) andThen { case (lcFhr, req) =>
       extendWithRequestIdLogCtx(implicit lc => {
-        val t0 = System.nanoTime
         logger.trace(s"Incoming request on ${req.uri}")
-        Timed
-          .future(metrics.daml.http_json_api.websocketRequest, lcFhr(lc))
-          .map(res => {
-            logger.trace(s"Processed request after ${System.nanoTime() - t0}ns")
-            res
-          })
+        lcFhr(lc)
       })
     }
   }
@@ -126,7 +119,7 @@ class WebsocketEndpoints(
       jwtPayload: domain.JwtPayload,
       req: WebSocketUpgrade,
       protocol: String,
-  )(implicit lc: LoggingContextOf[InstanceUUID with RequestID]): HttpResponse = {
+  )(implicit lc: LoggingContextOf[InstanceUUID with RequestID], metrics: Metrics): HttpResponse = {
     val handler: Flow[Message, Message, _] =
       webSocketService.transactionMessageHandler[A](jwt, jwtPayload)
     req.handleMessages(handler, Some(protocol))
