+++ ledger/metrics/BUILD.bazel
@@ -15,6 +15,7 @@ da_scala_library(
         "@maven//:com_typesafe_akka_akka_actor",
         "@maven//:com_typesafe_akka_akka_stream",
         "@maven//:org_scalaz_scalaz_core",
+        "@maven//:com_github_scopt_scopt",
     ],
     tags = ["maven_coordinates=com.daml:metrics:__VERSION__"],
     versioned_scala_deps = {
@@ -25,11 +26,22 @@ da_scala_library(
     ],
     runtime_deps = [],
     deps = [
+        "//ledger/ledger-resources",
         "//libs-scala/concurrent",
+        "//libs-scala/resources",
+        "//libs-scala/resources-akka",
+        "//libs-scala/resources-grpc",
         "@maven//:io_dropwizard_metrics_metrics_core",
+        "@maven//:io_dropwizard_metrics_metrics_graphite",
+        "@maven//:io_dropwizard_metrics_metrics_jmx",
         "@maven//:io_dropwizard_metrics_metrics_jvm",
+        "@maven//:io_netty_netty_transport",
         "@maven//:io_opentelemetry_opentelemetry_api",
         "@maven//:io_opentelemetry_opentelemetry_context",
+        "@maven//:io_prometheus_simpleclient",
+        "@maven//:io_prometheus_simpleclient_dropwizard",
+        "@maven//:io_prometheus_simpleclient_httpserver",
+        "@maven//:org_slf4j_slf4j_api",
     ],
 )
