+++ ledger/metrics/src/main/scala/com/daml/metrics/Metrics.scala
@@ -687,7 +687,7 @@ final class Metrics(val registry: MetricRegistry) {
       private val Prefix: MetricName = daml.Prefix :+ "http_json_api"
 
       val httpRequest: Timer = registry.timer(Prefix :+ "http_request")
-      val websocketRequest: Timer = registry.timer(Prefix :+ "websocket_request")
+      val websocketRequest: Counter = registry.counter(Prefix :+ "websocket_request")
     }
   }
 }
