+++ ledger/sandbox-common/BUILD.bazel
@@ -37,7 +37,6 @@ da_scala_library(
         "//ledger/ledger-api-auth",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-domain",
-        "//ledger/ledger-resources",
         "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-state",
@@ -45,13 +44,8 @@ da_scala_library(
         "//libs-scala/concurrent",
         "//libs-scala/contextualized-logging",
         "//libs-scala/ports",
-        "//libs-scala/resources",
-        "//libs-scala/resources-akka",
-        "//libs-scala/resources-grpc",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_auth0_java_jwt",
-        "@maven//:io_dropwizard_metrics_metrics_core",
-        "@maven//:io_dropwizard_metrics_metrics_jmx",
         "@maven//:io_netty_netty_handler",
     ],
 )
@@ -137,6 +131,7 @@ da_scala_library(
         "//ledger/ledger-api-domain",
         "//ledger/ledger-resources",
         "//ledger/ledger-resources:ledger-resources-test-lib",
+        "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-state",
         "//ledger/test-common",
